+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## XXX - 2021-XX
+## 1.7.1 - 2021-12-23
 - Add validateNameExpressions, getDomainNamePreviews methods to Domain
 
 ## 1.7.0 - 2021-12-02